+++ common/src/parsing/timeline.rs
@@ -15,6 +15,7 @@ pub struct Timeline<'s> {
     pub midi_instruments: BTreeMap<Cow<'s, str>, MidiInstrumentNumber>,
     pub csound_instruments: BTreeMap<Cow<'s, str>, CsoundInstrumentId<'s>>,
     pub csound_global_instruments: Vec<CsoundGlobalInstrument<'s>>,
+    pub csound_template: Option<Cow<'s, str>>,
     /// Least common multiple of time denominators, useful for computing ticks per beat. Callers
     /// should not count on 100% of denominators being a factor, but all denominators of note and
     /// duration values will be. This means you should take the numerator of the floor of the